+++ crates/project/src/dependencies.rs
@@ -254,10 +254,8 @@ impl DependencyVersionScheme {
         // dependency resolution
         match Self::try_from(spec)? {
             Self::Path { path: uri, version } => {
-                let workspace_path = workspace
-                    .source_file
-                    .as_ref()
-                    .map(|file| Path::new(file.content().uri().path()));
+                let workspace_path =
+                    workspace.source_file.as_ref().and_then(|file| file.content().uri().to_path());
                 if uri.scheme().is_none_or(|scheme| scheme == "file")
                     && let Some(workspace_path) = workspace_path.and_then(|p| p.canonicalize().ok())
                     && let Some(workspace_root) = workspace_path.parent()